modified log_getter to fully automate testing & graph generation

+++ s_log_getter.sh
@@ -1,9 +1,45 @@
 #!/bin/bash
+
+# 1. name server VM srv-32
+# 2. name client VM cli-32
+# 3. set up 3 net cards eth1...eth3 for each VM 
+# 4. add frac_digits(3) to syslog-ng options
+
+DBOXHOST=grandrew@alternet.homelinux.net # host to upload JSON logs to and parse them on
+DBOXHOST_PORT=10023
+VTRUNKD_L_ROOT=~/sandbox/vtrunkd
+VTRUNKD_V_ROOT=/home/user/sandbox/vtrunkd_test1
+LCNT=~/log_getter.counter
+
+VSRV_ETH1_IP=192.168.57.101
+VSRV_ETH2_IP=192.168.58.101
+VSRV_ETH3_IP=192.168.59.101
+
+VCLI_ETH1_IP=192.168.57.100
+VCLI_ETH2_IP=192.168.58.100
+VCLI_ETH3_IP=192.168.59.100
+
+NTP_SERVER="0.ubuntu.pool.ntp.org"
+
+if [ ! -f $LCNT ]; then i
+    RND=$(($RANDOM % 99))
+    echo -n "$RND"00 > $LCNT
+fi
+
 TEST="0"
-PREFIX="prefix_"
+EXEC="0"
+
+COUNT=$((`cat $LCNT`+1));
+echo -n $COUNT > $LCNT
+
+PREFIX="$COUNT"_
+
 while getopts :tp: OPTION
 do
  case $OPTION in
+ e) echo "Execute vtrunkd only"
+  EXEC="1"
+  ;;
  t) echo "Full speed test"
   TEST="1"
   ;;
@@ -16,30 +52,76 @@ do
  ;;
  esac
 done
+
+echo "Doing with prefix $PREFIX"
 echo "Starting..."
+echo "killall vtrunkd ... "
+ssh user@srv-32 "sudo killall -9 vtrunkd && sudo ipcrm -M 567888"
+ssh user@cli-32 "sudo killall -9 vtrunkd && sudo ipcrm -M 567888"
 echo "Clear syslog"
 ssh user@cli-32 "cat /dev/null | sudo tee /var/log/syslog"
 ssh user@srv-32 "cat /dev/null | sudo tee /var/log/syslog"
+echo "Copying vtrunkd sources ..."
+ssh user@cli-32 "mkdir -p $VTRUNKD_V_ROOT"
+ssh user@srv-32 "mkdir -p $VTRUNKD_V_ROOT"
+scp -r $VTRUNKD_L_ROOT/* user@srv-32:$VTRUNKD_V_ROOT/
+scp -r $VTRUNKD_L_ROOT/* user@cli-32:$VTRUNKD_V_ROOT/
+echo "Compiling vtrunkd ..."
+if ssh user@srv-32 "cd $VTRUNKD_V_ROOT; make clean; make"; then 
+    echo "OK"
+else
+    echo "Compile Error!"
+    exit 0;
+fi
+echo "Compiling ..."
+if ssh user@cli-32 "cd $VTRUNKD_V_ROOT; make clean; make"; then
+    echo "OK"
+else
+    echo "Compile Error!"
+    exit 0;
+fi
+
 echo "NTP sync..."
-ssh user@cli-32 "sudo ntpdate 192.168.0.101" &
+ssh user@cli-32 "sudo ntpdate $NTP_SERVER" &
 sleep 1
-ssh user@srv-32 "sudo ntpdate 192.168.0.101"
+ssh user@srv-32 "sudo ntpdate $NTP_SERVER"
+echo "Setting IP addresses..."
+if ssh user@srv-32 "sudo ifconfig eth1 $VSRV_ETH1_IP && sudo ifconfig eth2 $VSRV_ETH2_IP && sudo ifconfig eth3 $VSRV_ETH3_IP"; then
+    echo "OK"
+else 
+    echo "IP setup error"
+    exit 0
+fi
+if ssh user@cli-32 "sudo ifconfig eth1 $VCLI_ETH1_IP && sudo ifconfig eth2 $VCLI_ETH2_IP && sudo ifconfig eth3 $VCLI_ETH3_IP"; then
+    echo "OK"
+else 
+    echo "IP setup error"
+    exit 0
+fi
+echo "Applying emulation TC rules"
+ssh user@srv-32 "sudo $VTRUNKD_V_ROOT/test/srv_emulate_2.sh"
+
 echo "Starting server..."
-ssh user@srv-32 "sudo /home/user/sandbox/vtrunkd_test1/vtrunkd -s -f /home/user/sandbox/vtrunkd_test1/test/vtrunkd-srv.test.conf -P 5003"
+ssh user@srv-32 "sudo $VTRUNKD_V_ROOT/vtrunkd -s -f $VTRUNKD_V_ROOT/test/vtrunkd-srv.test.conf -P 5003"
 sleep 5
 echo "Starting client 1..."
-ssh user@cli-32 "sudo /home/user/sandbox/vtrunkd_test1/vtrunkd -f /home/user/sandbox/vtrunkd_test1/test/vtrunkd-cli.test.conf atest1 192.168.57.101 -P 5003"
+ssh user@cli-32 "sudo $VTRUNKD_V_ROOT/vtrunkd -f $VTRUNKD_V_ROOT/test/vtrunkd-cli.test.conf atest1 $VSRV_ETH1_IP -P 5003"
 echo "Starting client 2..."
-ssh user@cli-32 "sudo /home/user/sandbox/vtrunkd_test1/vtrunkd -f /home/user/sandbox/vtrunkd_test1/test/vtrunkd-cli.test.conf atest2 192.168.58.101 -P 5003"
+ssh user@cli-32 "sudo $VTRUNKD_V_ROOT/vtrunkd -f $VTRUNKD_V_ROOT/test/vtrunkd-cli.test.conf atest2 $VSRV_ETH2_IP -P 5003"
 sleep 8
 echo "Full started"
+if [ $EXEC = "1" ]; then
+    "Execute only!"
+    exit 0;
+fi
 echo "Worcking..."
-echo "time_starttransfer %{time_starttransfer} time_total %{time_total} speed_download %{speed_download}" | curl -m 150 --connect-timeout 4 http://10.200.1.31/u -o /dev/null -w @- > /tmp/${PREFIX}speed
-echo "" >>  /tmp/${PREFIX}speed
-ping -c 10 -q -a 10.200.1.31 | tail -3 >> /tmp/${PREFIX}speed
+echo "time_starttransfer %{time_starttransfer} time_total %{time_total} speed_download %{speed_download}" | ssh user@cli-32 "curl -m 150 --connect-timeout 4 http://10.200.1.31/u -o /dev/null -w @- > /tmp/${PREFIX}speed"
+ssh user@cli-32 'echo "" >>  /tmp/${PREFIX}speed'
+ssh user@cli-32 "ping -c 10 -q -a 10.200.1.31 | tail -3 >> /tmp/${PREFIX}speed"
 echo "killall vtrunkd"
 ssh user@srv-32 "sudo killall -9 vtrunkd && sudo ipcrm -M 567888"
 ssh user@cli-32 "sudo killall -9 vtrunkd && sudo ipcrm -M 567888"
+# NOT WORKING CODE -->>>>>>>>>>>>>>>
 if [ $TEST = "1" ]; then
  echo "Speed testing..."
  echo "time_starttransfer %{time_starttransfer} time_total %{time_total} speed_download %{speed_download}" | curl -m 150 --connect-timeout 4 http://192.168.57.101/u -o /dev/null -w @- > /tmp/${PREFIX}speed_eth1
@@ -62,6 +144,7 @@ if [ $TEST = "1" ]; then
  ping -c 10 -q -a 192.168.58.101 | tail -3 >> /tmp/${PREFIX}speed_eth2
 echo "efficiency factor - ${AG_EFF}% C_grow - ${C_GROW}% C_use - ${fdsfsdf}%" >> /tmp/${PREFIX}speed
 fi
+# <<<<<<<<<<<<<<-- END NOT WORKING CODE
 echo "Transfer syslogs"
 scp user@cli-32:/var/log/syslog /tmp/${PREFIX}syslog-cli
 scp user@srv-32:/var/log/syslog /tmp/${PREFIX}syslog-srv
@@ -76,8 +159,11 @@ grep "{\"p_" /tmp/${PREFIX}syslog-1_srv > /tmp/${PREFIX}syslog-1_srv_json
 grep "{\"p_" /tmp/${PREFIX}syslog-1_cli > /tmp/${PREFIX}syslog-1_cli_json
 grep "{\"p_" /tmp/${PREFIX}syslog-2_srv > /tmp/${PREFIX}syslog-2_srv_json
 grep "{\"p_" /tmp/${PREFIX}syslog-2_cli > /tmp/${PREFIX}syslog-2_cli_json
-scp /tmp/${PREFIX}* andrey@bonanza:~/sandbox/alarm_logs/
-rm /tmp/${PREFIX}syslog*
+echo "Uploading logs..."
+scp -P $DBOXHOST_PORT /tmp/${PREFIX}*json $DBOXHOST:~/Dropbox/alarm_logs/
+#rm /tmp/${PREFIX}syslog*
+echo "Drawing graphs"
+ssh -p $DBOXHOST_PORT $DBOXHOST "cd ~/Dropbox/alarm_logs/; python ./parse_json_fusion.py $COUNT"
 echo "Clear syslog"
 ssh user@cli-32 "cat /dev/null | sudo tee /var/log/syslog"
 ssh user@srv-32 "cat /dev/null | sudo tee /var/log/syslog"
